more norminette stuff

+++ ft_memmove.c
@@ -14,13 +14,13 @@
 
 void	*memmove(void *dest, const void *src, size_t n)
 {
-	char *srccpy;
-	char *destcpy;
-	char *buffer;
-	unsigned int i;
+	char	*srccpy;
+	char	*destcpy;
+	char	*buffer;
+	size_t	i;
 
-	srccpy = (char*)src;
-	destcpy = (char*)dest;
+	srccpy = (char *)src;
+	destcpy = (char *)dest;
 	i = 0;
 	while (i < n)
 	{
